fix(daemon): address reviewer findings on #1957 (dup detection, StatusHandler, heartbeat restore, proxy docs)
- Add handler-count assertion (EXPECTED=51) after registerHandlers to detect
  duplicate or missing registrations at startup
- Extract status handler into StatusHandler class (handlers/status.ts),
  removing serveHandlers as an IpcServer instance field
- Fix HEARTBEAT_INTERVAL_MS Object.defineProperty restore in spec to include
  writable:true so subsequent test mutations aren't silently no-ops
- Document static getter/setter proxies on IpcServer and EventStreamServer
- File #1961 (unbounded eventSubscribers Set) and #1962 (dispose streams)
